Rename tycho-embedder-api to tycho spi
Fix https://github.com/eclipse-tycho/tycho/issues/1696
