+++ pom.xml
@@ -507,7 +507,7 @@
 		<module>sisu-osgi</module>
 		<module>p2-maven-plugin</module>
 		<module>tycho-metadata-model</module>
-		<module>tycho-embedder-api</module>
+		<module>tycho-spi</module>
 		<module>tycho-core</module>
 		<module>tycho-testing-harness</module>
 		<module>tycho-compiler-jdt</module>
