+++ tycho-buildtimestamp-jgit/pom.xml
@@ -33,7 +33,7 @@
 		</dependency>
 		<dependency>
 			<groupId>org.eclipse.tycho</groupId>
-			<artifactId>tycho-embedder-api</artifactId>
+			<artifactId>tycho-spi</artifactId>
 			<version>${project.version}</version>
 		</dependency>
 		<dependency>
