+++ tycho-bundles/org.eclipse.tycho.embedder.shared/pom.xml
@@ -24,7 +24,7 @@
 	<packaging>jar</packaging>
 
 	<name>Tycho Embedder API Maven/OSGi Shared Classes</name>
-	<description>Classes shared between the tycho-embedder-api module and Tycho's OSGi bundles.</description>
+	<description>Classes shared between the tycho-spi module and Tycho's OSGi bundles.</description>
 	<dependencies>
 		<dependency>
 			<groupId>org.eclipse.platform</groupId>
