+++ tycho-compiler-plugin/src/main/java/org/eclipse/tycho/compiler/OsgiCompilerMojo.java
@@ -25,7 +25,7 @@ import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugins.annotations.LifecyclePhase;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.ResolutionScope;
-import org.eclipse.tycho.classpath.SourcepathEntry;
+import org.eclipse.tycho.SourcepathEntry;
 import org.eclipse.tycho.core.osgitools.project.BuildOutputJar;
 
 /**
