+++ tycho-compiler-plugin/src/main/java/org/eclipse/tycho/compiler/OsgiTestCompilerMojo.java
@@ -27,7 +27,7 @@ import org.apache.maven.plugins.annotations.LifecyclePhase;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.ResolutionScope;
 import org.eclipse.tycho.ReactorProject;
-import org.eclipse.tycho.classpath.SourcepathEntry;
+import org.eclipse.tycho.SourcepathEntry;
 import org.eclipse.tycho.core.dotClasspath.SourceFolderClasspathEntry;
 import org.eclipse.tycho.core.osgitools.DefaultReactorProject;
