+++ tycho-core/pom.xml
@@ -208,7 +208,7 @@
 		</dependency>
 		<dependency>
 			<groupId>org.eclipse.tycho</groupId>
-			<artifactId>tycho-embedder-api</artifactId>
+			<artifactId>tycho-spi</artifactId>
 			<version>${project.version}</version>
 		</dependency>
 		<dependency>
