+++ tycho-core/src/test/java/org/eclipse/tycho/core/test/TychoTest.java
@@ -22,9 +22,9 @@ import org.apache.maven.plugin.testing.SilentLog;
 import org.apache.maven.project.MavenProject;
 import org.codehaus.plexus.logging.Logger;
 import org.eclipse.tycho.ArtifactType;
+import org.eclipse.tycho.ClasspathEntry;
 import org.eclipse.tycho.DependencyArtifacts;
 import org.eclipse.tycho.TargetEnvironment;
-import org.eclipse.tycho.classpath.ClasspathEntry;
 import org.eclipse.tycho.core.TargetPlatformConfiguration;
 import org.eclipse.tycho.core.TychoProject;
 import org.eclipse.tycho.core.osgitools.DefaultReactorProject;
