+++ tycho-spi/.settings/org.eclipse.jdt.core.prefs
