+++ tycho-spi/.settings/org.eclipse.jdt.ui.prefs
