+++ tycho-spi/src/main/java/org/eclipse/tycho/classpath/ClasspathContributor.java
@@ -2,6 +2,7 @@ package org.eclipse.tycho.classpath;
 
 import java.util.List;
 
+import org.eclipse.tycho.ClasspathEntry;
 import org.eclipse.tycho.ReactorProject;
 
 public interface ClasspathContributor {
