+++ tycho-spi/src/main/java/org/eclipse/tycho/resolver/TychoResolver.java
