+++ tycho-surefire/tycho-surefire-plugin/src/main/java/org/eclipse/tycho/surefire/provider/impl/JUnit59WithVintageProvider.java
@@ -22,7 +22,7 @@ import java.util.Set;
 
 import org.apache.maven.model.Dependency;
 import org.codehaus.plexus.component.annotations.Component;
-import org.eclipse.tycho.classpath.ClasspathEntry;
+import org.eclipse.tycho.ClasspathEntry;
 import org.eclipse.tycho.surefire.provider.spi.TestFrameworkProvider;
 import org.osgi.framework.Version;
 import org.osgi.framework.VersionRange;
