+++ tycho-surefire/tycho-surefire-plugin/src/test/java/org/eclipse/tycho/surefire/provider/impl/AbstractJUnitProviderTest.java
@@ -20,8 +20,8 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.eclipse.tycho.ArtifactType;
+import org.eclipse.tycho.ClasspathEntry;
 import org.eclipse.tycho.DefaultArtifactKey;
-import org.eclipse.tycho.classpath.ClasspathEntry;
 import org.eclipse.tycho.core.osgitools.DefaultClasspathEntry;
 import org.junit.Before;
 import org.junit.Test;
